feat: move tree_sitter away
