+++ crates/config/src/lib.rs
@@ -42,8 +42,8 @@ pub fn from_yaml_string<'a, L: Language + Deserialize<'a>>(
 #[cfg(test)]
 mod test {
   use super::*;
-  use ast_grep_core::language::TSLanguage;
   use ast_grep_core::matcher::{Pattern, PatternBuilder, PatternError};
+  use ast_grep_core::tree_sitter::TSLanguage;
   use ast_grep_core::StrDoc;
   use ast_grep_core::{Language, LanguageExt};
   use std::path::Path;
