+++ crates/napi/src/napi_lang.rs
@@ -1,5 +1,5 @@
-use ast_grep_core::language::TSLanguage;
 use ast_grep_core::matcher::{Pattern, PatternBuilder, PatternError};
+use ast_grep_core::tree_sitter::TSLanguage;
 use ast_grep_core::LanguageExt;
 use ast_grep_dynamic::{CustomLang, DynamicLang};
 use ast_grep_language::{Language, SupportLang};
