+++ crates/pyo3/src/py_lang.rs
@@ -1,6 +1,6 @@
 use anyhow::Context;
-use ast_grep_core::language::TSLanguage;
 use ast_grep_core::matcher::{Pattern, PatternBuilder, PatternError};
+use ast_grep_core::tree_sitter::TSLanguage;
 use ast_grep_core::{LanguageExt, StrDoc};
 use ast_grep_dynamic::{CustomLang, DynamicLang};
 use ast_grep_language::{Language, SupportLang};
